+++ Controller/Controller/Controller.csproj
@@ -173,7 +173,6 @@
     <Compile Include="Helper\FileHelper.cs" />
     <Compile Include="Helper\ViewsHelper.cs" />
     <Compile Include="Helper\WindowGenerator.cs" />
-    <Compile Include="MainWindow\BlinkManager.cs" />
     <Compile Include="MainWindow\ChildController.cs" />
     <Compile Include="MainWindow\IterationManagerController.cs" />
     <Compile Include="MainWindow\MeasurementRoutine\MeasurementRoutineLoader.cs" />
